Plantalyze/requirements.txt
# conda 가상환경(Plantalyze) 사용 패키지
# python 버전 == 3.12.1
# 리눅스에서 사용불가한 패키지
# intel-openmp==2021.4.0
# mkl==2021.4.0
# pefile==2023.2.7
# pyinstaller==6.8.0
# pyinstaller-hooks-contrib==2024.7
# pywin32==306
# pywin32-ctypes==0.2.2
# tbb==2021.12.0
# tensorflow-intel==2.16.1
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fastjsonschema==2.20.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.4
fsspec==2024.5.0
gast==0.5.4
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==1.26.4
openai==1.33.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
tinycss2==1.3.0
torch==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.24
urllib3==2.2.1
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0